+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Convars for shuffling music & videos.
 - Convar for letting external resources handle the loadscreen shutdown.
 - Shutdown when spawnmanager isn't available, and `externalShutdown` is false.
+- Error log to easily see & copy errors when they happen in the NUI.
 
 ### Changed